[TBA-6.1] Support premium stuff

+++ src/types/animation.rs
@@ -38,7 +38,7 @@ pub struct Animation {
     pub mime_type: Option<Mime>,
 
     /// File size in bytes.
-    pub file_size: Option<u32>,
+    pub file_size: Option<u64>,
 }
 
 #[cfg(test)]
